Sync high precision CKKS bootstrapping

+++ src/poseidon/advance/homomorphic_mod.h
@@ -27,7 +27,7 @@ public:
 
     inline SineType type() const { return this->type_; }
 
-    inline int k() const { return this->k_; }
+    inline double k() const { return this->k_; }
 
     inline double q_div() const { return this->q_div_; }
 
@@ -45,8 +45,8 @@ public:
 
     inline double sqrt_2pi() const { return this->sqrt_2pi_; }
 
-    inline int32_t sine_poly_a() const { return sine_poly_a_; }
-    inline int32_t sine_poly_b() const { return sine_poly_b_; }
+    inline double sine_poly_a() const { return sine_poly_a_; }
+    inline double sine_poly_b() const { return sine_poly_b_; }
     inline const Polynomial &sine_poly() const { return this->sine_poly_; }
 
     inline const Polynomial &arcsine_poly() const { return this->arcsine_poly_; }
@@ -62,14 +62,14 @@ private:
     uint32_t level_start_;
     uint32_t log_message_ratio_;
     uint32_t double_angle_;
-    uint32_t k_;
+    double k_;
     double q_div_;
     double q_diff_;
     double sc_fac_;
     double sqrt_2pi_;
     Polynomial sine_poly_;
-    int32_t sine_poly_a_;
-    int32_t sine_poly_b_;
+    double sine_poly_a_;
+    double sine_poly_b_;
 
     Polynomial arcsine_poly_;
 };
